# perf(cq): optimize the fk code